Fix a tiny but serious mistake
This was more than likely causing the sub CPU to crash on detecting a main CPU error! The sub CPU's subsequent attempt to signal the main CPU was being mistreated by the main CPU as an acknowledgement. Add the wordram already being assigned to the main CPU, and you have a bug that was completely masked from view!